requirements.txt
# Django
django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9

# Celery & Redis
celery==5.3.6
redis==5.0.1

# Data Ingestion & Utilities
pandas==2.2.2
openpyxl==3.1.2
python-dotenv==1.0.1
gunicorn==22.0.0